feat: 3 Layered Architecture중 레포지토리 분리
